Implement mechanisms to allow omission of macro parameters.
When declaring a macro, the new keyword "optional" can now be placed in front of any identifier in the parameter list. When calling the macro, parameters flagged this way can be left empty, or undeclared identifiers passed to them. In both cases the corresponding parameter will be undeclared in the macro.
The new syntax "local(IDENTIFIER)" can be used to specifically access an identifier at the most local scope (i.e. the current macro, or the current file if used outside of a macro), ignoring any less local definitions; this can be used to reliably test for the presence of optional parameters using '#ifdef(local(IDENTIFIER))' or pass optional parameters on to other macros, without potential interference from global variables of the same name.