Extended tuple assignment syntax to allow for omission of right-hand side elements; extend array initializer syntax to allow for omission of elements.
Left-hand side elements in tuple assignments can now be prepended with the "optional" keyword. If an element is flagged this way, or left empty, the corresponding right-hand side element may be an uninitialized identifier or left empty. In this case, the respective left-hand side element remains unchanged (it is NOT undefined).
An array initializer may now be prepended with the "optional" keyword (e.g. "#declare A = array[10] optional { ... }"). In this case, each individual element of the initializer may be an uninitialized identifier or left empty.

Note: Due to the internal twists of the parser, "uninitialized identifier" does _not_ encompass uninitialized array elements. This limitation also applies to optional macro parameters.