Extended torus syntax to give more control over the spindle in a self-intersecting torus:
    `torus {` FLOAT `,` FLOAT [ `union` | `merge` | `difference` | `intersection` ] ... `}`

By default, or if the `union` keyword is specified, the spindle surface will be visible inside the torus, and its volume considered part of the object.
If the `merge` keyword is specified, the spindle surface will be invisible.
If the `difference` keyword is specified, the spindle volume will be considered outside the object.
If the `intersection` keyword is specified, _only_ the spindle will be visible and considered inside.

Node that with respect to the `interior_texture` mechanism, the spindle surface is always textured as if the spindle volume was inside and the bulk of the torus was outside, making it consistent with the texturing of the torus proper.