Fixed bug that could cause strange parse errors when including a file from a macro.
During a macro call, POV-Ray searches for the macro's `#end` by scanning for any `#` at the expected position. POV-Ray would continue to do this even while parsing a file included from the macro, potentially leading to directives in the included file being erroneously misidentified as the macro's proper `#end`, leading to failure to properly "step out of" of the included file, causing the parser to exhibit erratic behaviour. While this would ideally call for a more systematic solution to unify the mechanisms for tracking macro calls and includes, for now this has been fixed by adding a simple filename check, in order to minimize changes to the parser.