Added function-based user-defined camera.
The syntax is `camera { user_defined location { FUNCTION_TRIPLET } direction { FUNCTION_TRIPLET } CAMERA_MODIFIERS }`. Instead of `{ FUNCTION_TRIPLET }`, a constant vector can be specified, and the usual defaults apply. Each `FUNCTION_TRIPLET` is a set of three functions, all taking as input parameters the image x and y position ranging from -0.5 (left/bottom) to +0.5 (right/top), and acting together to return a ray origin and direction for a given pixel. A direction of <0,0,0> indicates that the pixel is to be left black.