Implemented tonemapping.
To use tonemapping, the user needs to supply one or more functions computing the effective output data from the raw data: One or three functions to manipulate the colour data, and optionally a second or fourth function to manipulate the transparency. By default, each function takes the corresponding channel's raw data as a single parameter, but the user can specify a different set of data to be passed to all functions.

The syntax is `global_settings { tonemapping { parameters { CHANNEL ... } FUNCTION ... } }`, where each `CHANNEL` may be any of `channel` (for the channel's raw data), `red`, `green`, `blue`, `filter` (for the named channel's raw data), `gray`, `grey` (for the raw grayscale value), 'x', 'y' (for the screen coordinate ranging from left/top = 0 to right/bottom = 1). each `FUNCTION` may take either the form `function { EXPRESSION }` or `function(IDENTIFIERS) { EXPRESSION }`, with the latter form allowing to specify the identifiers by which the chosen data channels will be referred to in EXPRESSION (defaulting to `x`, `y` and `z`).