Removing modulus op from DensityFilePattern::EvaluateRaw
There was previously a modulus operation in the switch() for every point
evaluated by DensityFilePattern::EvaluateRaw. It had this comment: TODO -
why the modulus operation??. This addition happened between 3.5 and
3.6.  Guessing perhaps trying to limit values to less than the 0-4095
eval "recommended" enum range limit on which the switch() was based.
Maybe some compiler hard coded to the recommendation, but elsewhere
plenty of switch() sets ups with int() so very unsure.
This update removes the modulus operation while adding stricter
enforcement to the parsing in Parser::Parse_Pattern and
Parser::Parse_PatternFunction. No change in behavior expected.  CPU
change across a set of tri-linear test cases was -1.38%. Values
seen by the switch() are more limited.