Adding image_map support to density.
A less risky alternative to adding pigment_map support to
density(#81-closed). This method also enables image slices to define a
density in a way other than the external image file slices to df3
methods. The densities would be defined with image files and those
would be used in a density_map.