Adding better pulling black_hole warp type 1.
The existing black_hole type 0 doesn't clamp properly at the surface of
the spherical region for the inward pulling black hole.

When inverted (pushing outward), clamping to an internal strength of 1.0
is enough because well never look further inward than the origin of the
spherical region for the point to push outward.

When pulling inward, clamping an internal strength of 1.0 OK for
effects, but it does not prevent the code from reaching outside the
spherical region for the point to pull.  The type below does and
provides for a more realistic and controllable inward pulling black_hole
by using the point to surface distance for the basis of movement.