Eliminate use of `jversion.h`.
We're now determining the LibJPEG version from the contents of `jpeglib.h` rather than `jversion.h`, as the latter does not constitute part of the official library interface as distributed in libjpeg-dev. As a side effect we can now stop creating a fake `jversion.h` as part of the Unix build process.