Fixing issues with bounding threshold option.
Related to ussues #59 and #180. When no bounding threshold is specified
centralizes centralizes the setting and reporting of the bounding
threshold to 3. Further, restores the clamping of values to 1+ which
existed in 3.6.

In deciding to apply this fix to the 3.71 release, this change was
rebased onto release/v3.7.1 and a new pull request create replacing
request #181 which was based upon a later master branch version.