Major overhaul of version number macro system in `version.h`.
Replaced the various redundant representations of the version number as authoritative sources of information with a set of integers, each one corresponding to a single numerical field of the version number. Any other representations of the version number are now automatically derived from these (plus the prerelease tag, if applicable).

Also added a mechanism that allows v3.8 binaries to run on top of a v3.7 host installation for now, so that we don't need any 3.8 installers yet.

Note that this commit comes with an updated Git pre-commit hook; to install it, copy `pre-commit` from `tools/git/hooks/` to `.git/hooks/` in your local repository.