Leaving mesh inside_vector untouched on transforms.
Formally, both the ray origin and the direction derived from the
inside_vector for the inside mesh test would see an inverse transform
into the mesh's coordinate space.  This would sometimes result in a
change in the inside_vector/ray direction with respect to the mesh.
Such direction changes are problematic given test ray directions roughly
perpendicular to a triangle's normal are noisy with respect to
ray-triangle intersections.