Fixes to scenes/advanced/pawns.pov (#455)
* Fix Z-fighting in pawns.pov

The chessboard of "one pine block and 32 little rosewood ones" created a solid in which both objects overlapped, leading to z-fighting (speckling) on the rendered output.  Replacing the large object with 32 small pine blocks corrects the issue.

* Replace `filter` with `transmit` in pawns.pov

pawns.pov was originally a 3.0 demo scene, where `filter` and `transmit` meant the same thing.  The behavior of `filter` was changed in 3.5, but old behavior could be gotten either through `#version` or by changing `filter` to `transmit` where needed.

Though the `#version` statement on this sample file got updated at some point, the `color_map` statements did not, leading to incorrect colors that differ from the original outputs.  This change returns the textures to their intended state.